Fixed multi-threading issue
If multiple instances of BsonParser were created in the same thread
but then accessed in multiple other threads, the same static buffers
were used which eventually made the reatUTF() method fail. Because static
buffers are used in readUTF() only, StaticBuffers.getInstance() can be
moved there. This solves the issue but has a slight performance impact.
However, since SoftReferences are used in StaticBuffers the performance
impact is neglectable.